+++ trunk/Simulation/src/main/java/sk/stuba/fiit/kvasnicka/qsimdatamodel/data/components/UsageStatistics.java
@@ -28,5 +28,5 @@ public interface UsageStatistics {
      *
      * @return
      */
-    int getUsage();
+    double getUsage();
 }
