+++ trunk/Simulation/src/main/java/sk/stuba/fiit/kvasnicka/qsimdatamodel/data/components/buffers/RxBuffer.java
@@ -158,7 +158,7 @@ public class RxBuffer implements UsageStatistics {
     }
 
     @Override
-    public int getUsage() {
+    public double getUsage() {
         return getNumberOfFragments();
     }
 }
